Fix compile time warning and possibly worse

+++ src/operations_liste.c
@@ -2292,7 +2292,7 @@ gboolean changement_taille_liste_ope ( GtkWidget *clist,
     gint col0, col1, col2, col3, col4, col5, col6;
 
     /* Yet another kludge to avoid breakage due to #392 fix (closes: #470) */
-    if ( compte != compte_courant_onglet )
+    if ( (gint) compte != compte_courant_onglet )
 	return TRUE;
 
     /* si la largeur de grisbi est < 700, on fait rien */
